requirements.txt
flask>=2.0.0
playwright>=1.30.0
openai>=1.0.0
python-dotenv>=0.19.0
pydantic>=2.0.0
requests>=2.26.0 
gunicorn>=23.0.0